Added playCard and playRound.
Under the WarHand constructor, added a name value, as well as a function
called playCard that returns the first card in the hand's cards array, and
runs the .show() function on it.  Then added a playRound function that plays
a card from each player's hand, compares them, and adds them to the winner's
hand.